detection module refactored
